threatintel/requirements.txt
Django==4.0.4
environs==9.5.0
psycopg2-binary==2.9.3
requests==2.27.1
djangorestframework==3.13.1
beautifulsoup4==4.11.1
confluent-kafka==1.8.2
stix2-elevator
stix2
flatdict
dagster
kafka-python
dagit
django-filter